+++ preload.js
@@ -66,7 +66,7 @@ contextBridge.exposeInMainWorld('electronAPI', {
 
   // 皮肤系统
   getAvailableSkins: () => ipcRenderer.invoke('get-available-skins'),
-  setCurrentSkin: (skinId) => ipcRenderer.send('set-current-skin', skinId),
+  setCurrentSkin: (skinId) => ipcRenderer.invoke('set-current-skin', skinId),
   onSwitchSkin: (callback) => {
     return subscribeIpc('switch-skin', (_event, skinId) => callback(skinId));
   },